+++ docs/src/changelog.md
@@ -20,6 +20,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 
 #### WWW
 - Update Web manifest
+- Update FontAwesome to 7.0.0
 - Log view rendering errors
 - Rework home view rendering
 - Ensure services are initialized before starting the server